site/www/requirements.txt
celery[redis]
Django==1.10.5
django-celery
django-extensions
django-material>=0.8.0
django-model-utils
django-jwt-auth
django-webpack-loader
docker-py
Jinja2
ansible
mysqlclient
psycopg2
django-markdown-deux
pygments
